refactor: rename auth_service to user_auth_service
