+++ bank/bank-client/src/routes/Logout.js
@@ -2,7 +2,7 @@ import { useContext } from 'react'
 import { Navigate } from 'react-router-dom';
 
 /* Services */
-import auth_service from '../services/auth_service';
+import user_auth_service from '../services/user_auth_service';
 
 /* Contexts */
 import userContext from '../contexts/userContext';
@@ -12,7 +12,7 @@ const Logout = () => {
     const user = userState.user;
     const setUser = userState.setUser;
 
-    auth_service.user_logout(user.id)
+    user_auth_service.user_logout(user.id)
         .then(() => {
             console.log("Successful Logout");
             setUser({});
